refactor: Register 파일 구조 재정리(#47) - Register 폴더 안에 Register.jsx, Register.css 올겼다 - 동의 항목 내용을 수정했다
