+++ src/routes/Register/Register.jsx
@@ -1,7 +1,7 @@
 import "./Register.css";
-import useRegister from "../hooks/useRegister";
-import checkedIcon from "../assets/icons/checked-icon.png";
-import uncheckedIcon from "../assets/icons/unchecked-icon.png";
+import useRegister from "./hooks/useRegister";
+import checkedIcon from "../../assets/icons/checked-icon.png";
+import uncheckedIcon from "../../assets/icons/unchecked-icon.png";
 
 function Register() {
   const { 
@@ -64,7 +64,7 @@ function Register() {
                 <label>학번</label>
                 <input
                   type="text"
-                  placeholder="2021111222"
+                  placeholder="5800000"
                   value={studentId}
                   onChange={handleStudentIdChange}
                 />
@@ -96,7 +96,7 @@ function Register() {
                 </label>
               </div>
               <div className="information-box-content">
-                <p>○○○은 ○○○서비스 회원가입, 고객상담 및 AS, 고지사항 전달 등을 위해 아래와 같이 정보를 수집합니다.</p>
+                <p>크무톡톡은 회원가입, 회원 관리 등을 위해 아래와 같이 개인정보를 수집·이용합니다.</p>
                 <table className="info-table">
                   <thead>
                     <tr>
@@ -107,21 +107,21 @@ function Register() {
                   </thead>
                   <tbody>
                     <tr>
-                      <td>회원 식별 및 회원제 서비스 제공</td>
-                      <td>아이디, 비밀번호</td>
+                      <td>회원 가입 시 재학생 인증</td>
+                      <td>이메일</td>
                       <td rowSpan="3">개인정보 보호법 제 15조 제 1항</td>
                     </tr>
                     <tr>
-                      <td>고객 상담 및 AS 관리</td>
-                      <td>전화번호</td>
+                      <td>학생 식별 및 관리자 답변 관리</td>
+                      <td>학번</td>
                     </tr>
                     <tr>
-                      <td>서비스 변경사항 및 고지사항 전달</td>
-                      <td>이메일</td>
+                      <td>사용자 식별</td>
+                      <td>쿠키</td>
                     </tr>
                   </tbody>
                 </table>
-                <p>귀하는 ○○○의 서비스 이용에 필요한 최소한의 개인정보 수집·이용에 동의하지 않을 수 있으나, 동의를 거부할 경우 회원제 서비스 이용이 불가능합니다.</p>
+                <p>귀하는 크무톡톡 서비스 이용에 필요한 최소한의 개인정보 수집·이용에 동의하지 않을 수 있으나, 동의를 거부할 경우 회원제 서비스 이용이 불가능합니다.</p>
               </div>
               <div className="btn-wrapper">
                 <button
